+++ src/examples/sizeBinding/index.ts
@@ -2,7 +2,7 @@ import {
   run,
   // DEV
   logFns,
-  mergeStates,
+  mergeComponents,
 } from '../../core'
 import { viewHandler } from '../../interfaces/view'
 import { styleHandler } from '../../groups/style'
@@ -29,7 +29,7 @@ let DEV = ENV === 'development'
   if (module.hot) {
     module.hot.accept('./Root', () => {
       let m = <any> require('./Root')
-      app.moduleAPI.reattach(m, mergeStates)
+      app.moduleAPI.reattach(m, mergeComponents)
     })
   }
 })()