+++ src/examples/sizeTask/index.ts
@@ -2,7 +2,7 @@ import {
   run,
   // DEV
   logFns,
-  mergeStates,
+  mergeComponents,
 } from '../../core'
 import { viewHandler } from '../../interfaces/view'
 import { sizeHandler } from '../../tasks/size'
@@ -33,7 +33,7 @@ let DEV = ENV === 'development'
   if (module.hot) {
     module.hot.accept('./Root', () => {
       let m = <any> require('./Root')
-      app.moduleAPI.reattach(m, mergeStates)
+      app.moduleAPI.reattach(m, mergeComponents)
     })
   }
 })()