+++ src/utils/reattach.spec.ts
@@ -1,7 +1,7 @@
 import { ComponentSpaceIndex } from '../core'
-import { mergeStates } from './reattach'
+import { mergeComponents } from './reattach'
 
-describe('mergeStates function should merge the states of the lastComponents and the context', () => {
+describe('mergeComponents function should merge the states of the lastComponents and the context', () => {
 
   let ctx: any = {
     id: '',
@@ -47,7 +47,7 @@ describe('mergeStates function should merge the states of the lastComponents and
     let newState = { a: 10, b: 0 }
     let components = createCompIndex(newState, state)
     let lastComponents = createCompIndex(modifiedState, state)
-    let result = mergeStates(<any> {}, components, lastComponents)
+    let result = mergeComponents(<any> {}, components, lastComponents)
     expect(result.Main.state).toEqual({ a: 10, b: 3 })
   })
 
@@ -57,7 +57,7 @@ describe('mergeStates function should merge the states of the lastComponents and
     let newState = 3
     let components = createCompIndex(newState, state)
     let lastComponents = createCompIndex(modifiedState, state)
-    let result = mergeStates(<any> {}, components, lastComponents)
+    let result = mergeComponents(<any> {}, components, lastComponents)
     expect(result.Main.state).toEqual(3)
   })
 
@@ -67,7 +67,7 @@ describe('mergeStates function should merge the states of the lastComponents and
     let newState = {}
     let components = createCompIndex(newState, state)
     let lastComponents = createCompIndex(modifiedState, state)
-    let result = mergeStates(<any> {}, components, lastComponents)
+    let result = mergeComponents(<any> {}, components, lastComponents)
     expect(result.Main.state).toEqual({ a: 1 })
   })
 
@@ -77,7 +77,7 @@ describe('mergeStates function should merge the states of the lastComponents and
     let newState = 0
     let components = createCompIndex(newState, state)
     let lastComponents = createCompIndex(modifiedState, state)
-    let result = mergeStates(<any> {}, components, lastComponents)
+    let result = mergeComponents(<any> {}, components, lastComponents)
     expect(result.Main.state).toEqual(2)
   })
 
@@ -89,7 +89,7 @@ describe('mergeStates function should merge the states of the lastComponents and
     let componentsAux = createCompIndex(123, 123)
     components.New = componentsAux.Main
     let lastComponents = createCompIndex(modifiedState, state)
-    let result = mergeStates(<any> {}, components, lastComponents)
+    let result = mergeComponents(<any> {}, components, lastComponents)
     expect(result.New.state).toEqual(123)
   })
 
@@ -111,7 +111,7 @@ describe('mergeStates function should merge the states of the lastComponents and
     }
     let components = createCompIndex(newState, state)
     let lastComponents = createCompIndex(modifiedState, state)
-    let result = mergeStates(<any> {}, components, lastComponents)
+    let result = mergeComponents(<any> {}, components, lastComponents)
     expect(result.Main.state).toEqual({
       list: {
         1: 'a',
@@ -143,7 +143,7 @@ describe('mergeStates function should merge the states of the lastComponents and
       ...createCompIndex(modifiedState, state, 'Main'),
       ...dynamic, // dynamic component
     }
-    let result = mergeStates(<any> {}, components, lastComponents)
+    let result = mergeComponents(<any> {}, components, lastComponents)
     // normal behavior
     expect(result.Main.state).toEqual({
       list: {
@@ -182,11 +182,11 @@ describe('mergeStates function should merge the states of the lastComponents and
       ...dynamic, // dynamic component
     }
     let error = (source, description) => {
-      expect(source).toEqual('mergeStates')
+      expect(source).toEqual('mergeComponents')
       expect(description).toEqual('there are no dynamic component definition of Main$other (defs) in Main')
       done()
     }
-    mergeStates(<any> { error }, components, lastComponents)
+    mergeComponents(<any> { error }, components, lastComponents)
   })
 
 })